Extract interface CookiePref to cookie-pref.ts
